move reducers/app.js to store/app.js
